// src/app/models/MaterialDTO.ts
export interface MaterialDTO {
  numero: string
  uf?: string
  codigo?: string
  lacre?: string
  recebidoLigado?: boolean
  bateria?: number
  telaFuncionando?: number
  recebidoModoAviao?: boolean
  recebidoBloqueado?: boolean
  senhaFornecida?: boolean
  senha?: string
  fotos?: {
    embalagem: string
    lacre: string
    detalhes: string
    simCards: string
    memoryCard: string
  }
  descricao?: string
  codigoEpol?: string
  estado_conservacao?: string
  aparencia_tela?: string
  funcionamento_tela?: string
  funcionamento_touch?: string
  funcionamento_botoes?: string
  funcionamento_conector_dados?: string
  outros_defeitos_observados?: string
  qtde_simcard?: number
  simcard1_operadora?: string
  simcard1_numero?: string
  simcard2_operadora?: string
  simcard2_numero?: string
  simcard3_operadora?: string
  simcard3_numero?: string
  qtde_memorycard?: number
  fabricante?: string
  modelo?: string
  imei1?: string
  imei2?: string
  serial?: string
  is_modo_aviao?: boolean
  is_simcard1_extracted?: boolean
  is_simcard2_extracted?: boolean
  is_simcard3_extracted?: boolean
  is_memorycard1_extracted?: boolean
  is_memorycard2_extracted?: boolean
  is_memorycard3_extracted?: boolean
  is_inseyets_extracting?: boolean
  inseyets_laped_machine?: string
  is_physical_analyzer_opening?: boolean
  physical_analyzer_laped_machine?: string
  whatsapp_physical_analyzer?: string
  dados_usuario?: string
  is_iped_opening?: boolean
  is_iped_ok?: boolean
  is_zipping?: boolean
  is_zip_ok?: boolean
  is_zip_moving?: boolean
}